Improvement: Move plugin validator to validators.py.
